게시물 수정, 삭제시 권한체크

+++ src/main/java/com/jhs/exam/exam2/http/controller/UsrArticleController.java
@@ -52,12 +52,19 @@ public class UsrArticleController extends Controller {
 		}
 
 		Article article = articleService.getForPrintArticleById(id);
-		
-		if ( article == null ) {
+
+		if (article == null) {
 			rq.historyBack(Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 			return;
 		}
-		
+
+		ResultData actorCanDeleteRd = articleService.actorCanDelete(rq.getLoginedMember(), article);
+
+		if (actorCanDeleteRd.isFail()) {
+			rq.historyBack(actorCanDeleteRd.getMsg());
+			return;
+		}
+
 		articleService.delete(id);
 
 		rq.replace(Ut.f("%d번 게시물을 삭제하였습니다.", id), redirectUri);
@@ -72,8 +79,8 @@ public class UsrArticleController extends Controller {
 		}
 
 		Article article = articleService.getForPrintArticleById(id);
-		
-		if ( article == null ) {
+
+		if (article == null) {
 			rq.historyBack(Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 			return;
 		}
@@ -117,13 +124,13 @@ public class UsrArticleController extends Controller {
 	private void actionShowWrite(Rq rq) {
 		rq.jsp("usr/article/write");
 	}
-	
+
 	private void actionDoModify(Rq rq) {
 		int id = rq.getIntParam("id", 0);
 		String title = rq.getParam("title", "");
 		String body = rq.getParam("body", "");
 		String redirectUri = rq.getParam("redirectUri", Ut.f("../article/detail?id=%d", id));
-		
+
 		if (id == 0) {
 			rq.historyBack("id를 입력해주세요.");
 			return;
@@ -136,6 +143,20 @@ public class UsrArticleController extends Controller {
 
 		if (body.length() == 0) {
 			rq.historyBack("body를 입력해주세요.");
+			return;
+		}
+
+		Article article = articleService.getForPrintArticleById(id);
+
+		if (article == null) {
+			rq.historyBack(Ut.f("%d번 게시물이 존재하지 않습니다.", id));
+			return;
+		}
+
+		ResultData actorCanModifyRd = articleService.actorCanModify(rq.getLoginedMember(), article);
+
+		if (actorCanModifyRd.isFail()) {
+			rq.historyBack(actorCanModifyRd.getMsg());
 			return;
 		}
 
@@ -153,8 +174,15 @@ public class UsrArticleController extends Controller {
 		}
 
 		Article article = articleService.getForPrintArticleById(id);
-		
-		if ( article == null ) {
+
+		ResultData actorCanModifyRd = articleService.actorCanModify(rq.getLoginedMember(), article);
+
+		if (actorCanModifyRd.isFail()) {
+			rq.historyBack(actorCanModifyRd.getMsg());
+			return;
+		}
+
+		if (article == null) {
 			rq.historyBack(Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 			return;
 		}
